@@ -9,15 +9,16 @@
     <version>0.3.0</version>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>20</maven.compiler.source>
+        <maven.compiler.target>20</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <grpc.version>1.49.1</grpc.version>
-        <protobuf.version>3.21.6</protobuf.version>
+        <grpc.version>1.61.0</grpc.version>
+        <protobuf.version>3.25.2</protobuf.version>
         <protoc.version>3.19.2</protoc.version>
-        <kafka.version>3.2.3</kafka.version>
-        <mockito.version>4.8.0</mockito.version>
-        <junit.version>5.9.1</junit.version>
+        <kafka.version>3.6.1</kafka.version>
+        <mockito.version>5.9.0</mockito.version>
+        <junit.version>5.10.1</junit.version>
+        <logback.version>1.4.14</logback.version>
     </properties>
 
     <dependencyManagement>
