@@ -29,7 +29,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class DebeziumPgSourceIT extends BasePostgresIT {
+class DebeziumPgSourceIT extends BasePostgresIT {
     @Override
     protected Map<String, String> configMap() {
         return Map.of(
@@ -40,7 +40,8 @@ public class DebeziumPgSourceIT extends BasePostgresIT {
             "database.password", PASSWORD,
             "database.dbname", "meroxadb",
             "database.server.name", "test-server",
-            "table.include.list", "public.employees"
+            "table.include.list", "public.employees",
+            "topic.prefix", "test-topic"
         );
     }
 
