@@ -16,12 +16,23 @@
 
 package io.conduit;
 
+import java.util.Arrays;
+import java.util.Map;
+import java.util.NoSuchElementException;
+
+import io.conduit.grpc.Specifier;
+import io.conduit.grpc.Specifier.Parameter.Validation;
 import io.conduit.grpc.Specifier.Specify.Request;
 import io.conduit.grpc.Specifier.Specify.Response;
 import io.grpc.stub.StreamObserver;
+import lombok.SneakyThrows;
+import org.apache.kafka.common.config.ConfigDef;
+import org.apache.kafka.connect.sink.SinkConnector;
+import org.apache.kafka.connect.source.SourceConnector;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -33,8 +44,9 @@ import static org.mockito.Mockito.verify;
 
 class SpecifierServiceTest {
     @Test
-    public void testSpecify() {
+    void testSpecify() {
         var observer = mock(StreamObserver.class);
+
         new SpecifierService().specify(
             Request.newBuilder().build(),
             observer
@@ -61,8 +73,114 @@ class SpecifierServiceTest {
 
         assertNotNull(response.getSourceParamsMap());
         assertFalse(response.getSourceParamsMap().isEmpty());
+        verifyWrapperClass(
+            new String[]{
+                "io.aiven.connect.jdbc.JdbcSourceConnector",
+                "io.debezium.connector.postgresql.PostgresConnector"
+            },
+            response.getSourceParamsMap()
+        );
+        verifySourceParams(
+            "io.aiven.connect.jdbc.JdbcSourceConnector",
+            response.getSourceParamsMap()
+        );
+        verifySourceParams(
+            "io.debezium.connector.postgresql.PostgresConnector",
+            response.getSourceParamsMap()
+        );
 
         assertNotNull(response.getDestinationParamsMap());
         assertFalse(response.getDestinationParamsMap().isEmpty());
+        verifyWrapperClass(
+            new String[]{
+                "io.aiven.connect.jdbc.JdbcSinkConnector",
+                "com.snowflake.kafka.connector.SnowflakeSinkConnector"
+            },
+            response.getDestinationParamsMap()
+        );
+        verifyDestinationParams(
+            "io.aiven.connect.jdbc.JdbcSinkConnector",
+            response.getDestinationParamsMap()
+        );
+        verifyDestinationParams(
+            "com.snowflake.kafka.connector.SnowflakeSinkConnector",
+            response.getDestinationParamsMap()
+        );
+    }
+
+    @SneakyThrows
+    private void verifySourceParams(String clazz, Map<String, Specifier.Parameter> paramsMap) {
+        SourceConnector connector = (SourceConnector) newInstance(clazz);
+        connector.config()
+            .configKeys()
+            .forEach((keyName, key) -> verifyParam(clazz, key, paramsMap));
+    }
+
+    @SneakyThrows
+    private void verifyDestinationParams(String clazz, Map<String, Specifier.Parameter> paramsMap) {
+        SinkConnector connector = (SinkConnector) newInstance(clazz);
+        connector.config()
+            .configKeys()
+            .forEach((keyName, key) -> verifyParam(clazz, key, paramsMap));
+    }
+
+    private void verifyParam(String clazz, ConfigDef.ConfigKey key, Map<String, Specifier.Parameter> paramsMap) {
+        assertTrue(
+            paramsMap.containsKey(clazz + "." + key.name),
+            "expected params to contain: " + key.name
+        );
+        Specifier.Parameter param = paramsMap.get(clazz + "." + key.name);
+        // Description
+        assertEquals(key.documentation, param.getDescription());
+        // Required and default value
+        boolean required = ConfigDef.NO_DEFAULT_VALUE.equals(key.defaultValue);
+        if (required) {
+            assertTrue(
+                param.getValidationsList()
+                    .stream()
+                    .anyMatch(v -> Validation.Type.TYPE_REQUIRED.equals(v.getType()))
+            );
+            assertEquals("", param.getDefault());
+        } else {
+            assertTrue(
+                param.getValidationsList()
+                    .stream()
+                    .noneMatch(v -> Validation.Type.TYPE_REQUIRED.equals(v.getType()))
+            );
+            assertEquals(String.valueOf(key.defaultValue), param.getDefault());
+        }
+    }
+
+    @SneakyThrows
+    private Object newInstance(String className) {
+        Class<?> clazz = Class.forName(className);
+        return Arrays.stream(clazz.getConstructors())
+            .filter(c -> c.getParameterCount() == 0)
+            .findFirst()
+            // get() also throws NoSuchElementException, but here:
+            // 1. we send the caller more details
+            // 2. we also adhere to a best practice (not calling get() without checking if value is present)
+            .orElseThrow(() -> new NoSuchElementException("no parameterless constructor for " + className))
+            .newInstance();
+    }
+
+    private void verifyWrapperClass(String[] expected, Map<String, Specifier.Parameter> params) {
+        var validations = params.get("wrapper.connector.class").getValidationsList();
+        assertNotNull(validations);
+        assertEquals(2, validations.size());
+        var required = validations.stream()
+            .filter(v -> v.getType() == Validation.Type.TYPE_REQUIRED)
+            .findFirst();
+        assertTrue(required.isPresent());
+
+
+        var values = validations.stream()
+            .filter(v -> v.getType() == Validation.Type.TYPE_INCLUSION)
+            .findFirst();
+        assertTrue(values.isPresent());
+        var actual = values.get().getValue().split(",");
+        Arrays.sort(actual);
+        Arrays.sort(expected);
+        assertArrayEquals(expected, actual);
     }
 }
